etl/dags/requirements.txt
apache-airflow==1.10.10
google-cloud-storage==1.27.0
requests==2.23.0
